Created Form Assignment
Utilized HTML to create the structure of a form in a html file.
Utilized CSS external file linked to the HTML file to style the form element through use of classes.
Utilized JS file linked to HTML file to attach functions to the form alerting user that all fields must be filled entirely.